backend/requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
curlify==2.2.1
facebook-business==15.0.0
fastapi==0.88.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
multidict==6.0.2
pycountry==22.3.5
pydantic==1.10.2
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.1
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
urllib3==1.26.13
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4
yarl==1.8.1
